finance_bot_work/requirements.txt
python-telegram-bot==20.8
apscheduler==3.10.4
sqlalchemy==2.0.30
aiosqlite==0.18.0
bcrypt==4.1.3
python-binance==1.0.17
pybit==5.3.0
tinkoff-investments==0.2.0b110
aiohttp==3.9.3